Final Build Improved/backend/requirements.txt
Flask
google-generativeai
python-dotenv
wikipedia
requests